refactor(cert): relocate test-only composition doubles to test assembly
Move ControlledCompositionProposer, RecordedCompositionGeneratorModel,
DamageHealthCompositionProposals, CompositionProposalRecorder, and
RecordedCompositionProposal to Nexo.Tests.Infrastructure.Certification.Doubles.
Inline batch-entry replay into CompositionAcceptanceRateMeasurer; production
Infrastructure references no test-only types.
