+++ src/Nexo.Tests.Infrastructure/Tests/Certification/CompositionProposerDogfoodTests.cs
@@ -3,6 +3,7 @@ using Nexo.Core.Application.Certification.Models;
 using Nexo.Core.Application.Certification.Ports;
 using Nexo.Infrastructure.Certification;
 using Nexo.Infrastructure.Certification.Composition;
+using Nexo.Tests.Infrastructure.Certification.Doubles;
 using Nexo.Tests.Infrastructure.Certification.Dogfood;
 using Nexo.Tests.Infrastructure.Certification.Fixtures;
 using Xunit;
