+++ src/Nexo.Tests.Infrastructure/Tests/Certification/RealModelCompositionProposerDogfoodTests.cs
@@ -1,6 +1,7 @@
 using FluentAssertions;
 using Nexo.Core.Application.Certification.Models;
 using Nexo.Infrastructure.Certification.Composition;
+using Nexo.Tests.Infrastructure.Certification.Doubles;
 using Nexo.Tests.Infrastructure.Certification.Dogfood;
 using Nexo.Tests.Infrastructure.Certification.Fixtures;
 using Xunit;
